backend/requirements.txt
flask
pymongo
flask-cors
flask-jwt-extended
python-dotenv
werkzeug
